dd_icd10 {WHOmortality}	R Documentation

ICD-10 Mortality Data Dictionary

Description

Data dictionary for the ICD-10 codes used in the mortality table, extracted from Table 8 in the documentation Word document "Documentation_1May2019.doc"

Usage

data(dd_icd10)

Format

A data.frame object containing the column names and descriptions for the ICD-10 code table "mort"

code

code used in the mortality table

detailed

detailed ICD-10 code list

cause

general cause of death

Source

WHO Mortality Database

Examples

data(dd_icd10)
head(dd_icd10)

[Package WHOmortality version 0.0.0.9 Index]