dd_mort {WHOmortality}	R Documentation

WHO Mortality Data Dictionary

Description

Data dictionary for the mortality raw files extracted from the documentation Word document "Documentation_1May2019.doc"

Usage

data(dd_mort)

Format

A data.frame object containing the column names and descriptions for the mortality data tables "Morticd10_part1" and "Morticd10_part2"

column_name

the column name used for the WHO mortality table

content

the description of the column name

type

expected data type for the column

Source

WHO Mortality Database

Examples

data(dd_mort)
head(dd_mort)

[Package WHOmortality version 0.0.0.9 Index]