download_who {WHOmortality}	R Documentation

Download WHO Mortality Data

Description

This function downloads the raw WHO mortality data based on the links on the webpage https://www.who.int/healthinfo/statistics/mortality_rawdata/en/

Usage

download_who(dest, data = c("docs", "mort", "pop", "ccode", "all"))

Arguments

dest
a character file path of the destintation directory to download the .zip files

data
to specify which files to download, takes values of "docs", "mort", "pop", "ccode", "all". Default value downloads the documentation files

Details

The data parameter allows the user to download a specific set of data files. The docs downloads the documentation files associated with the raw WHO mortality data, this includes the "documentation.zip", "notes.zip", and "availability.zip" files The mort downloads the two ICD-10 mortality .zip files, "Morticd10_part1.zip" and "Morticd10_part2.zip" The pop downloads the WHO population data, "pop.zip" The ccode downloads the country code table "country_codes.zip" The all downloads all the abovementioned .zip files at once

Value

the same value as the download.file

Examples

## Not run: 
download_who(dest = getwd(), data = "ccode")

## End(Not run)

[Package WHOmortality version 0.0.0.9 Index]