extract_who {WHOmortality}	R Documentation

Extract WHO Mortality Data

Description

Extract WHO Mortality Data

Usage

extract_who(path, dest)

Arguments

path
a character file path where the downloaded WHO .zip files exist

dest
a character destination file path to extract, default value will be path

Value

a named logical vector of the files that were successfully extracted

Examples

## Not run: 
download_who(dest = getwd(), data = "ccode")
download_who(dest = getwd(), data = "docs")
extract_who(path = getwd(), dest = getwd(), convert = TRUE)

## End(Not run)

[Package WHOmortality version 0.0.0.9 Index]