proc_who_mort {WHOmortality}	R Documentation

Process Raw WHO Mortality Data

Description

This function processes the original WHO mortality csv file to correct data types for each column.

Usage

proc_who_mort(x)

Arguments

x
a data.frame of the original csv WHO mortality table, this is the output of import_who

Value

a data.frame object with modified column data types

Examples

## Not run: 
mort <- import_who(path = getwd(), data = "mort")
mort <- proc_who_mort(mort)

## End(Not run)

[Package WHOmortality version 0.0.0.9 Index]