snakecase_cols {WHOmortality}	R Documentation

Convert Column Names to Lower Snake Case

Description

This function converts all column names of input data.frame to lower snake case.

Usage

snakecase_cols(x)

Arguments

x
a data.frame of the original csv WHO mortality table, this is the output of import_who

Value

a data.frame with snake case column names

Examples

## Not run: 
ccodes <- import_who(path = getwd(), data = "ccode")
names(ccodes)
ccodes <- snakecase_cols(ccodes)
names(ccodes)

## End(Not run)

[Package WHOmortality version 0.0.0.9 Index]